fix: move registry type to enum fix: pr comments
