+++ src/github/workflows.ts
@@ -2,6 +2,7 @@
 import decamelize = require('decamelize');
 
 import { Component } from '../component';
+import { kebabCaseKeys } from '../util';
 import { YamlFile } from '../yaml';
 import { GitHub } from './github';
 
@@ -69,28 +70,6 @@ function snakeCaseKeys<T = unknown>(obj: T): T {
       v = snakeCaseKeys(v);
     }
     result[decamelize(k)] = v;
-  }
-  return result as any;
-}
-
-export function kebabCaseKeys<T = unknown>(obj: T, recursive = true): T {
-  if (typeof obj !== 'object' || obj == null) {
-    return obj;
-  }
-
-  if (Array.isArray(obj)) {
-    if (recursive) {
-      obj = obj.map((v) => kebabCaseKeys(v, recursive)) as any;
-    }
-    return obj;
-  }
-
-  const result: Record<string, unknown> = {};
-  for (let [k, v] of Object.entries(obj)) {
-    if (recursive) {
-      v = kebabCaseKeys(v, recursive);
-    }
-    result[decamelize(k).replace(/_/mg, '-')] = v;
   }
   return result as any;
 }
